Conditionally render join mission button
