+++ src/component/Missions.js
@@ -1,8 +1,7 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import DisplayMission from './DisplayMission';
 import styles from '../styles/Missions.module.css';
-import { getMission } from '../features/missions/missionSlice';
+import { getMission, joinMission } from '../features/missions/missionSlice';
 
 function Missions() {
   const { missions } = useSelector((store) => store.missions);
@@ -10,6 +9,12 @@ function Missions() {
   useEffect(() => {
     dispatch(getMission());
   }, [dispatch]);
+
+  function displayJoinMission(currentState) {
+    return currentState ? 'Leave Mission' : 'Join Mission';
+  }
+  const handleJoinMission = (id) => dispatch(joinMission(id));
+
   return (
     <div className={styles.missionsContainer}>
       <div className={styles.missionsHeader}>
@@ -24,13 +29,48 @@ function Missions() {
         </div>
         <div />
       </div>
-      {missions.map((mision) => (
-        <DisplayMission
-          key={mision.mission_id}
-          id={mision.mission_id}
-          name={mision.mission_name}
-          description={mision.description}
-        />
+      {missions.map((mission) => (
+        <div
+          className={styles.missionItems}
+          key={mission.mission_id}
+        >
+          <div className={styles.missionHead}>
+            <h3>{mission.mission_name}</h3>
+          </div>
+          <div className={styles.missionInfo}>
+            <p>{mission.description}</p>
+          </div>
+          <div className={styles.buttons}>
+            {mission.reserved ? (
+              <button
+                type="button"
+                className={styles.activeMemberBtn}
+              >
+                Active Member
+              </button>
+            ) : (
+              <button
+                type="button"
+                className={styles.memberBtn}
+              >
+                NOT A MEMBER
+              </button>
+            )}
+          </div>
+          <div className={styles.buttons}>
+            <button
+              onClick={() => handleJoinMission(mission.mission_id)}
+              type="button"
+              className={
+                mission.reserved
+                  ? styles.leaveMissionBtn
+                  : styles.joinMissionBtn
+              }
+            >
+              {displayJoinMission(mission.reserved)}
+            </button>
+          </div>
+        </div>
       ))}
     </div>
   );
